initial move to FileService interface
