+++ Backend/src/main/java/idealab/api/controller/DropboxController.java
@@ -2,7 +2,7 @@ package idealab.api.controller;
 
 import idealab.api.dto.request.DropBoxFilePathRequest;
 import idealab.api.dto.response.PrintJobResponse;
-import idealab.api.service.FileService;
+import idealab.api.service.FileServiceImpl;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,15 +12,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/dropbox")
 public class DropboxController {
 
-    private final FileService fileService;
+    private final FileServiceImpl fileServiceImpl;
 
-    public DropboxController(FileService fileService) {
-        this.fileService = fileService;
+    public DropboxController(FileServiceImpl fileServiceImpl) {
+        this.fileServiceImpl = fileServiceImpl;
     }
 
     @PutMapping("/file-path")
     public PrintJobResponse updateFilePath(@RequestBody DropBoxFilePathRequest request) {
-        return fileService.updateDropboxPath(request);
+        return fileServiceImpl.updateFilePath(request);
     }
 
 }
