+++ Backend/src/main/java/idealab/configurations/GoogleDriveConfig.java
@@ -1,0 +1,5 @@
+package idealab.configurations;
+
+
+public class GoogleDriveConfig {
+}
