+++ Backend/src/test/java/idealab/api/operations/PrintJobOperationsTest.java
@@ -7,8 +7,9 @@ import idealab.api.exception.IdeaLabApiException;
 import idealab.api.model.Queue;
 import idealab.api.model.*;
 import idealab.api.repositories.*;
-import idealab.api.service.FileService;
+import idealab.api.service.FileServiceImpl;
 import idealab.api.service.EmailHashUtil;
+import idealab.api.service.FileService;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -39,7 +40,7 @@ public class PrintJobOperationsTest {
     PrintJobRepo printJobRepo;
 
     @Mock
-    private FileService fileService;
+    private FileService fileServiceImpl;
 
     @Mock
     private ColorTypeRepo colorTypeRepo;
@@ -56,7 +57,7 @@ public class PrintJobOperationsTest {
     @Before
     public void setup() {
         operations = new PrintJobOperations(
-                fileService,
+                fileServiceImpl,
                 printJobRepo,
                 colorTypeRepo,
                 customerInfoRepo,
@@ -343,7 +344,7 @@ public class PrintJobOperationsTest {
         when(colorTypeRepo.findByColor(any())).thenReturn(color);
         when(employeeRepo.findEmployeeByUsername(any())).thenReturn(e);
         when(printJobRepo.save(any())).thenReturn(printJob);
-        when(fileService.uploadDropboxFile(anyLong(), any())).thenReturn(data);
+        when(fileServiceImpl.uploadFile(anyLong(), any())).thenReturn(data);
 
         PrintJobResponse opResponse = operations.newPrintJob(request, mockPrincipal);
         assert(opResponse.equals(response));
@@ -412,7 +413,7 @@ public class PrintJobOperationsTest {
         when(colorTypeRepo.findByColor(any())).thenReturn(color);
         when(employeeRepo.findEmployeeByUsername(any())).thenReturn(e);
         when(printJobRepo.save(any())).thenReturn(printJob);
-        when(fileService.uploadDropboxFile(anyLong(), any())).thenReturn(data);
+        when(fileServiceImpl.uploadFile(anyLong(), any())).thenReturn(data);
 
         PrintJobResponse opResponse = operations.newPrintJob(request, mockPrincipal);
 
@@ -551,7 +552,7 @@ public class PrintJobOperationsTest {
         data.put("sharableLink", "http://testlink.com");
 
         when(printJobRepo.findPrintJobById(printJob.getId())).thenReturn(printJob);
-        when(fileService.updateDropboxFile(printJob, request.getFile())).thenReturn(data);
+        when(fileServiceImpl.updateFile(printJob, request.getFile())).thenReturn(data);
         when(printJobRepo.save(printJob)).thenReturn(printJob);
 
         PrintJobResponse opResponse = operations.updateModel(printJob.getId(), request);
@@ -587,7 +588,7 @@ public class PrintJobOperationsTest {
         printJob.setId(999);
 
         when(printJobRepo.findPrintJobById(printJob.getId())).thenReturn(printJob);
-        when(fileService.updateDropboxFile(printJob, request.getFile())).thenThrow(new IdeaLabApiException(DROPBOX_UPLOAD_FILE_ERROR));
+        when(fileServiceImpl.updateFile(printJob, request.getFile())).thenThrow(new IdeaLabApiException(DROPBOX_UPLOAD_FILE_ERROR));
 
         operations.updateModel(printJob.getId(), request);
     }
@@ -603,7 +604,7 @@ public class PrintJobOperationsTest {
         printJob.setId(999);
 
         when(printJobRepo.findPrintJobById(printJob.getId())).thenReturn(printJob);
-        doNothing().when(fileService).deleteDropboxFile(printJob.getFilePath());
+        doNothing().when(fileServiceImpl).deleteFile(printJob.getFilePath());
         when(printJobRepo.save(printJob)).thenReturn(printJob);
 
         GenericResponse opResponse = operations.deleteModel(printJob.getId());
